Split settings into /settings (account) and /organization
Org config was sharing a page with account config while depending on
useCurrentOrg, so switching orgs silently redrew half of /settings. Now:

- /settings: your name, email, dark mode, delete account.
- /organization: org name, id, QR defaults, delete organization. New
  sidebar entry between Members and Billing, and reserved in
  RESERVED_SLUGS + robots.txt.

Skeletons split to match; e2e that renamed or deleted an org through
/settings now go through /organization.
